+++ src/main/java/net/lopymine/ip/config/particle/ParticleConfig.java
@@ -7,7 +7,7 @@ import java.util.*;
 import lombok.*;
 import net.lopymine.ip.element.*;
 import net.lopymine.ip.spawner.context.ParticleSpawnContext;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.lopymine.mossylib.utils.CodecUtils;
 import static net.lopymine.mossylib.utils.CodecUtils.option;
 
@@ -21,13 +21,13 @@ public class ParticleConfig {
 				return either.right().orElseGet(() -> either.left().map(DynamicParticleSizes::fromStatic).orElse(null));
 			}, Either::right);
 
-	public static final Codec<ResourceLocation> TEXTURES_CODEC = ResourceLocation.CODEC.xmap((id) -> {
+	public static final Codec<Identifier> TEXTURES_CODEC = Identifier.CODEC.xmap((id) -> {
 		if (id.getPath().endsWith(".png")) {
 			String path = id.getPath();
 			String i = id.getNamespace();
 			String s = path.substring(0, path.length() - 4);
 			//? if >=1.21 {
-			return ResourceLocation.fromNamespaceAndPath(i, s);
+			return Identifier.fromNamespaceAndPath(i, s);
 			//?} else {
 			/*return ResourceLocation.tryBuild(i, s);
 			 *///?}
@@ -49,7 +49,7 @@ public class ParticleConfig {
 	private ParticleAnimationType animationType;
 	private double animationSpeed;
 	private DynamicParticleSizes size;
-	private ArrayList<ResourceLocation> textures;
+	private ArrayList<Identifier> textures;
 	private HashSet<ParticleHolder> holders;
 	private ParticlePhysics physics;