+++ src/main/java/net/lopymine/ip/texture/LoopParticleTextureProvider.java
@@ -2,7 +2,7 @@ package net.lopymine.ip.texture;
 
 import java.util.*;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 
 public class LoopParticleTextureProvider extends StretchParticleTextureProvider {