+++ src/main/java/net/lopymine/ip/texture/RandomParticleTextureProvider.java
@@ -3,7 +3,7 @@ package net.lopymine.ip.texture;
 import java.util.List;
 import lombok.*;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.util.RandomSource;
 import org.jetbrains.annotations.Nullable;